Trait aimc_hal::axes::Axes

pub trait Axes {
    fn set_target_velocity(&mut self, axis_number: usize, velocity: Velocity);

    fn velocity(&self, axis_number: usize) -> Option<Velocity>;
}

A driver for controlling axis motion using velocity control.

Required methods

fn set_target_velocity(&mut self, axis_number: usize, velocity: Velocity)

Tell the specified axis to move at a desired velocity.

fn velocity(&self, axis_number: usize) -> Option<Velocity>

Get the actual velocity a particular axis is moving at.

Loading content...